requirements.txt
requests
emails
python-dotenv